add styles for product details and move product details to separate file
